apply CORS middleware before others
